+++ native/OutboundSalesiOS/OutboundSalesiOS/OutboundSalesiOSApp.swift
@@ -1,0 +1,11 @@
+import OutboundSalesNative
+import SwiftUI
+
+@main
+struct OutboundSalesiOSApp: App {
+    var body: some Scene {
+        WindowGroup {
+            OutboundSalesRootView()
+        }
+    }
+}